requirements.txt
requests
pandas
tqdm
numpy
openai
tiktoken
numba
scikit-learn
google-generativeai